+++ src/Workflowish/Item/ChildItems.tsx
@@ -24,7 +24,7 @@ export const ChildItems = (props: {
             id: "",
             children:[],
             collapsed: true,
-            searchHighlight: "NONE"
+            searchHighlight: []
         }]
     }